ENH: First pass for the internationalization.
Creation of the macro to generate/update the .ts and .qm files.

Add globals Cmake variables to recover the directories paths of the .qm
if slicer is not installed - Slicer_QM_OUTPUT_DIRS -, or recover
the directory path if slicer is installed - Slicer_QM_DIR -.

Add global CMake variable Slicer_LANGUAGE to set the languages wanted.

2 new functions in the main :
 - First, check if slicer is installed, and recover the directory(ies)
from the cmake variable corresponding.
 - recovered all the .qm files from a given directory, and load the
translation into the application.